playground/pyproject.toml
[project]
name = "agentic-playground"
version = "0.1.0"
description = "Interactive playground for testing agentic-primitives with live OTel events"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Neural", email = "neural@agentparadise.dev" }
]
keywords = ["agent", "claude", "opentelemetry", "playground", "testing"]

dependencies = [
    # Local packages (editable installs via workspace)
    "agentic-adapters",
    "agentic-isolation",
    "agentic-security",
    # CLI and display
    "rich>=13.0",
    "typer>=0.9",
    # Configuration
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
]

[project.scripts]
playground = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
agentic-adapters = { path = "../lib/python/agentic_adapters", editable = true }
agentic-isolation = { path = "../lib/python/agentic_isolation", editable = true }
agentic-security = { path = "../lib/python/agentic_security", editable = true }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
